Move tarball and umbrella handlers into their own files
